% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustcurv_surv.R
\name{clustcurv_surv}
\alias{clustcurv_surv}
\title{Clustering survival curves}
\usage{
clustcurv_surv(time, status, fac, kvector = NULL, kbin = 50, nboot = 100,
  algorithm = "kmeans", alpha = 0.05, cluster = FALSE, ncores = NULL,
  seed = NULL)
}
\arguments{
\item{time}{Survival time.}

\item{status}{Censoring indicator of the survival time of the process; 0 if
the total time is censored and 1 otherwise.}

\item{fac}{Categoriacl variable indicating the population to which
the subject belongs}

\item{kvector}{A vector specifying the number of groups of curves to be
checking.}

\item{kbin}{Size of the grid over which the survival functions
are to be estimated.}

\item{algorithm}{A character string specifying which clustering algorithm is used,
i.e., k-means(\code{"kmeans"}) or k-medians (\code{"kmedians"}).}

\item{alpha}{Seed to be used in the procedure.}

\item{cluster}{A logical value. If  \code{TRUE} (default), the
testing procedure is  parallelized. Note that there are cases
(e.g., a low number of bootstrap repetitions) that R will gain in
performance through serial computation. R takes time to distribute tasks
across the processors also it will need time for binding them all together
later on. Therefore, if the time for distributing and gathering pieces
together is greater than the time need for single-thread computing, it does
not worth parallelize.}

\item{ncores}{An integer value specifying the number of cores to be used
in the parallelized procedure. If \code{NULL} (default), the number of cores
to be used is equal to the number of cores of the machine - 1.}

\item{seed}{Seed to be used in the procedure.}
}
\value{
A list containing the following items:
 \item{table}{A data frame containing the null hypothesis tested, the values
 of the test statistics and the obtained pvalues.}
 \item{levels}{Original levels of the variable \code{fac}.}
 \item{cluster}{A vector of integers (from 1:k) indicating the cluster to
 which each curve is allocated.}
 \item{centers}{An object of class \code{survfit} containing the centroids
 (mean of the curves pertaining to the same group).}
 \item{curves}{An object of class \code{survfit} containing the survival
 curves for each population.}
}
\description{
Function for grouping survival curves based on the k-means or
k-medians algorithm. It returns the number of groups and the assignement.
}
\examples{
library(clustcurv)
library(survival)
data(veteran)

res <- clustcurv_surv(time = veteran$time, status = veteran$status,
fac = veteran$celltype, algorithm = "kmeans")

res <- clustcurv_surv(colonCSm$time, status = colonCSm$status,
fac = colonCSm$nodes, nboot = 20)


}
\author{
Marta Sestelo, Nora M. Villanueva.
}
